Refactor Http clients - first draft.
