+++ bravado/mapping/http_client.py
@@ -1,0 +1,106 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+#
+# Copyright (c) 2013, Digium, Inc.
+# Copyright (c) 2014, Yelp, Inc.
+#
+
+import urlparse
+"""HTTP client abstractions.
+"""
+
+APP_FORM = 'application/x-www-form-urlencoded'
+APP_JSON = 'application/json'
+MULT_FORM = 'multipart/form-data'
+
+
+class HttpClient(object):
+    """Interface for a minimal HTTP client.
+    """
+
+    def set_basic_auth(self, host, username, password):
+        """Configures client to use HTTP Basic authentication.
+
+        :param host: Hostname to limit authentication to.
+        :param username: Username
+        :param password: Password
+        """
+        raise NotImplementedError(
+            u"%s: Method not implemented", self.__class__.__name__)
+
+    def set_api_key(self, host, api_key, param_name=u'api_key'):
+        """Configures client to use api_key authentication.
+
+        The api_key is added to every query parameter sent.
+
+        :param host: Hostname to limit authentication to.
+        :param api_key: Value for api_key.
+        :param param_name: Parameter name to use in query string.
+        """
+        raise NotImplementedError(
+            u"%s: Method not implemented", self.__class__.__name__)
+
+    def request(self, request_params, op):
+        """
+        :param request_params: complete request data.
+        :type request_params: dict
+
+        :returns: The client's request object
+        """
+        raise NotImplementedError(
+            u"%s: Method not implemented", self.__class__.__name__)
+
+    def __repr__(self):
+        return "{0}()".format(type(self))
+
+
+class Authenticator(object):
+    """Authenticates requests.
+
+    :param host: Host to authenticate for.
+    """
+
+    def __init__(self, host):
+        self.host = host
+
+    def __repr__(self):
+        return u"%s(%s)" % (self.__class__.__name__, self.host)
+
+    def matches(self, url):
+        """Returns true if this authenticator applies to the given url.
+
+        :param url: URL to check.
+        :return: True if matches host, port and scheme, False otherwise.
+        """
+        split = urlparse.urlsplit(url)
+        return self.host == split.hostname
+
+    def apply(self, request):
+        """Apply authentication to a request.
+
+        :param request: Request to add authentication information to.
+        """
+        raise NotImplementedError(u"%s: Method not implemented",
+                                  self.__class__.__name__)
+
+
+# noinspection PyDocstring
+class ApiKeyAuthenticator(Authenticator):
+    """?api_key authenticator.
+
+    This authenticator adds a query parameter to specify an API key.
+
+    :param host: Host to authenticate for.
+    :param api_key: API key.
+    :param param_name: Query parameter specifying the API key.
+    """
+
+    def __init__(self, host, api_key, param_name=u'api_key'):
+        super(ApiKeyAuthenticator, self).__init__(host)
+        self.param_name = param_name
+        self.api_key = api_key
+
+    def apply(self, request):
+        request.params[self.param_name] = self.api_key
+        return request
